Let WebKit parse with clang on windows with -std=c++11 https://bugs.webkit.org/show_bug.cgi?id=85398
Reviewed by Ryosuke Niwa.

See http://trac.webkit.org/changeset/85945 for background on the line
I'm changing. With clang, __GXX_EXPERIMENTLAL_CXX0X__ is set in c++11
mode, but MSVC's c++ library doesn't have a tr1/memory header. So also
check for __GLIBCXX__, like it's done in the rest of this file.

* wtf/TypeTraits.h:


Identifier: 102971@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115893 268f45cc-cd09-0410-ab3c-d52691b4dbfc