[EFL] media/video-poster-blocked-by-willsendrequest.html is flaky https://bugs.webkit.org/show_bug.cgi?id=85609
Patch by Christophe Dumez <christophe.dumez@intel.com> on 2012-05-07
Reviewed by Simon Hausmann.

Flakiness was caused by the fact that
media/crash-closing-page-with-media-as-plugin-fallback.html and
media/video-poster-blocked-by-willsendrequest.html use the same PNG
file as poster attribute in the video tag. The PNG file gets cached
when the first test is executed and this somehow causes the second
test to fail.

The flakiness was addressed by clearing the memory and page caches
between tests.

* DumpRenderTree/efl/DumpRenderTreeChrome.cpp:
(DumpRenderTreeChrome::resetDefaultsToConsistentValues):

Identifier: 103359@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116296 268f45cc-cd09-0410-ab3c-d52691b4dbfc