[Chromium] Fix the position of an RTL resizer https://bugs.webkit.org/show_bug.cgi?id=80640
Reviewed by Tony Chang.

When rendering an RTL resizer, my r110073 renders it at the same position as the
one used for rendering a LTR resizer. Unfortunately, this code renders the RTL
resizer image at an incorrect position on Mac and Linux. This change uses the
correct formula to calculate the position of an RTL resizer.

Source/WebCore:

Test: platform/chromium/scrollbars/rtl-resizer-position.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::drawPlatformResizerImage):

LayoutTests:

* platform/chromium/platform/chromium/scrollbars: Added.
* platform/chromium/platform/chromium/scrollbars/rtl-resizer-position-expected.png: Added.
* platform/chromium/platform/chromium/scrollbars/rtl-resizer-position-expected.txt: Added.
* platform/chromium/scrollbars/rtl-resizer-position.html: Added.


Identifier: 103486@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116435 268f45cc-cd09-0410-ab3c-d52691b4dbfc