Fix the SOFT_LINK_STAGED_FRAMEWORK_OPTIONAL macro so it passes the full path to dlopen.
dyld only considers libraries in the versioned framework path if their install name
matches the library that it is attempting to load. The path we were passing to
dlopen lacked the Versions/A component of the path so dyld did not recognize that
we wanted it to use the staged version if it is newer.

<rdar://problem/11406517>

Reviewed by Mark Rowe.

Source/WebCore:

* platform/mac/SoftLinking.h: Have SOFT_LINK_STAGED_FRAMEWORK_OPTIONAL take the
framework version as an argument and use it when constructing the path to dlopen.

Source/WebKit/mac:

* WebCoreSupport/WebInspectorClient.mm: Pass A to SOFT_LINK_STAGED_FRAMEWORK_OPTIONAL for the version.

Source/WebKit2:

* UIProcess/mac/WebInspectorProxyMac.mm: Pass A to SOFT_LINK_STAGED_FRAMEWORK_OPTIONAL for the version.
* WebProcess/WebPage/mac/WebInspectorMac.mm: Ditto.

Identifier: 103499@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116449 268f45cc-cd09-0410-ab3c-d52691b4dbfc