[CMake] FindGStreamer: Fix the build with static WebCore. https://bugs.webkit.org/show_bug.cgi?id=85930
Reviewed by Daniel Bates.

.:

Building WebCore statically was failing because files in
WebCore/platform/gstreamer when ENABLE_VIDEO was set required
gstreamer-base, which was not being linked to after r116453.

Fix that by looking for gstreamer-base, requiring and linking
against it if GStreamer is used.

* Source/cmake/FindGStreamer.cmake:

Source/WebCore:

No new tests, build fix.

Building WebCore statically was failing because files in
WebCore/platform/gstreamer when ENABLE_VIDEO was set required
gstreamer-base, which was not being linked to after r116453.

Fix that by looking for gstreamer-base, requiring and linking
against it if GStreamer is used.

* PlatformEfl.cmake: Link to GSTREAMER_LIBRARIES and
GSTREAMER_BASE_LIBRARIES, and include GSTREAMER_INCLUDE_DIRS and
GSTREAMER_BASE_INCLUDE_DIRS.

Identifier: 103517@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116468 268f45cc-cd09-0410-ab3c-d52691b4dbfc