<rdar://problem/10972577> and https://bugs.webkit.org/show_bug.cgi?id=80170 Contents of noscript elements turned into strings in WebArchives
Source/WebCore:

Reviewed by Andy Estes.

There's a much deeper question about how innerHTML of <noscript> is expected to work in
both a scripting and non-scripting environment that we should pursue separately.

But for webarchives, we can solve this by filtering out the <noscript> elements completely
if scripting is enabled.

Test: webarchive/ignore-noscript-if-scripting-enabled.html

* WebCore.exp.in:

Add arguments to createMarkup and MarkupAccumulator methods to pass a Vector of QualifiedNames
that should be filtered from the resulting markup:
* editing/MarkupAccumulator.cpp:
(WebCore::MarkupAccumulator::serializeNodes):
(WebCore::MarkupAccumulator::serializeNodesWithNamespaces):
* editing/MarkupAccumulator.h:
* editing/markup.cpp:
(WebCore::createMarkup):
* editing/markup.h:

If scripting is enabled, add the noscriptTag to the tag names to filter:
* loader/archive/cf/LegacyWebArchive.cpp:
(WebCore::LegacyWebArchive::create):

LayoutTests:

Reviewed by Andy Estes.

* webarchive/ignore-noscript-if-scripting-enabled-expected.txt: Added.
* webarchive/ignore-noscript-if-scripting-enabled.html: Added.

Identifier: 103726@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@116685 268f45cc-cd09-0410-ab3c-d52691b4dbfc